Add js modules, refactor function and update README file
