+++ README.md
@@ -1,2 +1,57 @@
-# todo-app
-Minimalist HTML TODO list app, ES6 compliant and bundled with webpack
+# To Do list: list structure ✔️!
+
+Projects aims:
+
+* Set up a new project with webpack that is based on the webpack exercise you have already completed.
+* Create an index.html file and write your HTML markup here. Create an empty To Do List placeholder (<div> or <ul> element). The index.html file must be set as a template using the HTML Webpack Plugin.
+* Create an index.js file and set an array of some simple to do tasks (array of objects).
+* Write a function to iterate over the tasks array and populate an HTML list item element for each task.
+* On page load render the dynamically created list of tasks in the dedicated placeholder. The list should appear in order of the index values for each task.
+* Create a style.css and set rules for the To Do List. CSS must be loaded by Webpack Style/CSS Loader. Your list should be a clone of the part of the minimalist project captured in the video below.
+
+## Built With
+
+- HTML
+- JavaScript
+- CSS
+-Webpack
+
+## Prerequisites
+
+### Setup
+
+- Set up a new repository and prepare it for development using best practices (e.g. linters).
+- Create a new branch according to the GitHub flow rules.
+- Add a .gitignore file.
+- Set up a linter for HTML and CSS and make sure that any files that are not relevant are ignored by git.
+
+- Add descriptive README file to your project
+
+### Install
+
+For instructions of how to install the required dependencies, please visit:
+
+- https://github.com/microverseinc/linters-config
+
+Not familiar with node & npm, please visit:
+
+- https://docs.npmjs.com/downloading-and-installing-node-js-and-npm
+
+## Authors
+
+:bust_in_silhouette: Jose Ramon
+
+- GitHub: @jr-cast
+- Twitter: @josercastanos
+- LinkedIn: https://www.linkedin.com/in/jr-cast
+
+
+## Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the issues page.
+
+## Show your support
+
+Give a :star: if you like this project!
