+++ src/index.html
@@ -14,7 +14,7 @@
 
 <body>
   <div class="wrapper">
-    <ul id="todolist">
+    <div id="todolist">
       <div id="heading">
         <h3>Today's To Do</h3> <i class="fa-solid fa-rotate"></i>
       </div>
@@ -22,7 +22,7 @@
         <input id="input" type="text" placeholder="Add to your list...">
         <i class="fa-solid fa-floppy-disk"></i>
       </form>
-    </ul>
+    </div>
     <div id="clear"><a href="#clear">Clear all completed</a></div>
   </div>
 </body>
