+++ src/index.js
@@ -1,31 +1,17 @@
-import _ from 'lodash';
 import './style.css';
-
-const todos = [{
-  description: 'Go to the Gym',
-  completed: false,
-  index: 0
-}, {
-  description: 'Take the dog out',
-  completed: false,
-  index: 1
-}, {
-  description: 'Play videogames',
-  completed: false,
-  index: 2
-}];
+import todos from './modules/todosObject.js';
 
 const listItems = () => {
-  for (let i = 0; i < todos.length; i++) {
+  for (let i = 0; i < todos.length; i += 1) {
     const checkBox = document.createElement('input');
     checkBox.type = 'checkbox';
-    const listItem = document.createElement('li');
-    listItem.innerHTML = todos[i].description;
-    listItem.id = `list${i}`;
-    listItem.prepend(checkBox);
+    const item = document.createElement('div');
+    item.innerHTML = todos[i].description;
+    item.classList.toggle('item');
+    item.prepend(checkBox);
     const todo = document.getElementById('todolist');
-    todo.append(listItem);
+    todo.append(item);
   }
-}
+};
 
 listItems();
