+++ src/modules/todosObject.js
@@ -1,0 +1,15 @@
+const todos = [{
+  description: 'Go to the Gym',
+  completed: false,
+  index: 0,
+}, {
+  description: 'Take the dog out',
+  completed: false,
+  index: 1,
+}, {
+  description: 'Play videogames',
+  completed: false,
+  index: 2,
+}];
+
+export default todos;
