+++ src/style.css
@@ -48,7 +48,7 @@ body {
   margin-right: 1.1em;
 }
 
-li {
+.item {
   width: 90%;
   padding: 1em;
   margin: 0%;
@@ -84,5 +84,5 @@ input:placeholder-shown {
 }
 
 .wrapper {
-  box-shadow: 0 0 25px 0 rgba(0, 0, 0, 0.150);
+  box-shadow: 0 0 25px 0 rgba(0, 0, 0, 0.15);
 }
